Get rid of debug logging
I was testing a hypothesis that turned out to be partially false (at least in the case of our Storybook story). Here’s a timeline of what I hypothesized was happening
1. Drawer close is triggered and calls a callback that the application is looking for
2. The application issues a state change, which
    - Stops rendering the drawer
    - Tells the drawer-toggle component that the drawer has closed
3. Drawer-toggle component responds to the state change and calls focus on its button element
4. The drawer component unmounts, which calls the dialogNode.close() method, triggering a response from the browser which then tries to move focus to what it thinks should now have focus

It turns out with these debug logs that the unmount was happening after the manual call to focus, but the browser wasn't errantly changing the focused element as I had guessed. This was not tested, however, in the application in which this problem was reported