Testing out the same strategy we use in the current doc site
We're currently having [issues re-hydrating conditional sections](https://github.com/CMSgov/design-system/pull/2752) of content in our Gatsby site. The theme comes from query parameters, but the static html is always rendered with the core theme, so it has trouble reconciling the different version of the dom when another theme is selected.

It appears that we have similar issues rehydrating in Astro